backend/requirements.txt
Django==4.2.18
djangorestframework==3.15.2
mysqlclient>=2.0
python-dotenv>=0.15
django-cors-headers==4.6.0
django-environ==0.12.0

